+++ app/internal/modules/catalog/entity/admin-api/category_image.go
@@ -1,5 +1,0 @@
-package entity
-
-const (
-	ModelNameCategoryImage = "admin-api.Catalog.CategoryImage"
-)