+++ app/internal/modules/catalog/entity/public-api/category.go
@@ -16,7 +16,8 @@ type (
 	}
 
 	CategoryParams struct {
-		Filter CategoryListFilter
+		LanguageID uint16
+		Filter     CategoryListFilter
 	}
 
 	CategoryListFilter struct {