+++ app/internal/modules/catalog/usecase/public-api/category_lang.go
@@ -33,7 +33,7 @@ func (uc *CategoryLangDecorator) GetList(ctx context.Context, params entity.Cate
 		return nil, 0, err
 	}
 
-	if dict := uc.getDict(ctx); dict != nil {
+	if dict := uc.getDict(ctx, params.LanguageID); dict != nil {
 		for i := range items {
 			items[i].Caption = dict.ItemByID(int(items[i].ID)).Attr("caption", items[i].Caption)
 		}
@@ -42,20 +42,20 @@ func (uc *CategoryLangDecorator) GetList(ctx context.Context, params entity.Cate
 	return items, total, nil
 }
 
-func (uc *CategoryLangDecorator) GetItem(ctx context.Context, id mrtype.KeyInt32) (*entity.Category, error) {
-	item, err := uc.service.GetItem(ctx, id)
+func (uc *CategoryLangDecorator) GetItem(ctx context.Context, id mrtype.KeyInt32, languageID uint16) (*entity.Category, error) {
+	item, err := uc.service.GetItem(ctx, id, languageID)
 
 	if err != nil {
 		return nil, err
 	}
 
-	item.Caption = uc.getDict(ctx).ItemByID(int(item.ID)).Attr("caption", item.Caption)
+	item.Caption = uc.getDict(ctx, languageID).ItemByID(int(item.ID)).Attr("caption", item.Caption)
 
 	return item, nil
 }
 
-func (uc *CategoryLangDecorator) getDict(ctx context.Context) *mrlang.Dictionary {
-	dict, err := uc.dict.ByLangID(mrctx.Locale(ctx).LangID())
+func (uc *CategoryLangDecorator) getDict(ctx context.Context, languageID uint16) *mrlang.Dictionary {
+	dict, err := uc.dict.ByLangID(languageID)
 
 	if err != nil {
 		mrctx.Logger(ctx).Warn(err)