+++ migrations/000001_init.down.sql
@@ -4,6 +4,4 @@ DROP TABLE gs_catalog.products;
 DROP TABLE gs_catalog.trademarks;
 DROP TABLE gs_catalog.categories;
 
-DROP TYPE gs_catalog.item_status;
-
 DROP SCHEMA gs_catalog;